+++ src/Sinch/ApiError.cs
@@ -1,7 +1,8 @@
-﻿using System.Collections.Generic;
-#if NET6_0_OR_GREATER
+﻿#if NET6_0_OR_GREATER
 using System.Text.Json.Nodes;
+using System.Collections.Generic;
 #endif
+
 
 namespace Sinch
 {
