+++ src/Sinch/Conversation/Messages/Message/AppMessage.cs
@@ -1,5 +1,4 @@
 ﻿using System.Text;
-using System.Text.Json.Serialization;
 
 namespace Sinch.Conversation.Messages.Message
 {
