+++ src/Sinch/SMS/DeliveryReport.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Runtime.Serialization;
-using System.Text.Json;
+﻿using System.Runtime.Serialization;
 using System.Text.Json.Serialization;
 using Sinch.Core;
