+++ src/Sinch/SMS/DeliveryReports/DeliveryReportStatus.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Text.Json;
-using System.Text.Json.Serialization;
+﻿using System.Text.Json.Serialization;
 using Sinch.Core;
 
 namespace Sinch.SMS.DeliveryReports
