+++ src/Sinch/SMS/DeliveryReports/Get/Request.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 using Sinch.Core;
 
 namespace Sinch.SMS.DeliveryReports.Get
