+++ tests/Sinch.Tests/Conversation/SendMessageTests.cs
@@ -8,7 +8,6 @@ using System.Threading.Tasks;
 using FluentAssertions;
 using Newtonsoft.Json;
 using RichardSzalay.MockHttp;
-using Sinch.Conversation;
 using Sinch.Conversation.Messages;
 using Sinch.Conversation.Messages.Message;
 using Sinch.Conversation.Messages.Send;
