+++ tests/Sinch.Tests/TestBase.cs
@@ -19,7 +19,7 @@ namespace Sinch.Tests
         protected TestBase()
         {
             var httpClient = new HttpClient(HttpMessageHandlerMock);
-            _tokenManager.GetToken(false).Returns(Token);
+            _tokenManager.GetToken().Returns(Token);
             HttpCamelCase = new Http(_tokenManager, httpClient, null, JsonNamingPolicy.CamelCase);
             HttpSnakeCase = new Http(_tokenManager, httpClient, null, SnakeCaseNamingPolicy.Instance);
         }
